refactor: relocate check_content_access to epigraph-db shared layer (A3 §7.5 prep)
