+++ crates/epigraph-db/src/lib.rs
@@ -50,11 +50,15 @@
 //!
 //! All migrations are in `/migrations/` and should be run with `sqlx migrate run`.
 
+pub mod access_control;
 pub mod errors;
 pub mod pool;
 pub mod repos;
 
 // Re-export primary types
+pub use access_control::{
+    batch_check_content_access, check_content_access, ContentAccess, COARSE_EDGE_TYPES,
+};
 pub use errors::DbError;
 pub use pool::{create_pool, create_pool_from_options, create_pool_with_options};
 pub use repos::{
